% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_desc.R
\name{cervan_desc}
\alias{cervan_desc}
\title{Add description}
\usage{
cervan_desc(name_pkg = pkg_name(), title_pkg, desc_pkg,
  mail_pro = TRUE, repo_git = NULL)
}
\arguments{
\item{name_pkg}{Name of the package}

\item{title_pkg}{Title of the package}

\item{desc_pkg}{Desc of the package}

\item{mail_pro}{mail pro ou perso}

\item{repo_git}{URL repos gil}
}
\description{
Add description
}
