% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_training.R
\name{find_your_training}
\alias{find_your_training}
\title{Find the goog formation}
\usage{
find_your_training(name_training, token)
}
\arguments{
\item{name_training}{name of training}

\item{token}{token to toggl}
}
\description{
Find the goog formation
}
