% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_training.R
\name{day_of_training}
\alias{day_of_training}
\title{Add Training}
\usage{
day_of_training(name_tranning, day, token = Sys.getenv("TOGGL_TOKEN"))
}
\arguments{
\item{name_tranning}{name of training}

\item{day}{which day}

\item{token}{token toggl}
}
\description{
Add Training
}
